Write transformed definitions to the top of the file that the macro was expanded in instead of the file the macro was defined in.
Create a function for generating the name of a transformed macro. Include the filename that the macro came from in the transformed name so that if the expansion came from a macro that was defined in a header file that was included in multiple files, name conflcts won't arise when linking multiple files that have been transformed and all contained expansions of that macro.