Rework the transformation proof to be much cleaner.
We no longer separate the lemmas that the proof uses into separate files, and defined them declaritively instead of proving them manually as we did before.

We also introduce a new concept: R-value independence. This is essential to the proof. We believe that side-effect freedom in parameters guarantees this for macro evaluation stores and function evaluation stores, and have an informal proof sketch in mind, but have not formally proven it yet.