Fix an issue where attributes were not properly replaced after deduplicating
Instead of counting the number of raw deduplications, we now count the number of unique deduplications. A deduplication is unique if it occurs inside a definition of a function in the original program, or inside a canonical transformed definition.

I believe that because of the way we now handle header-defined macros, the number of unique deduped invocations should equal the number of unique deduped decls/definitions