Remove use of visitors.
Inline much of the deduplication logic into the consumer.

Define a method for printing debug messages if verbosity is turned on in deduplicator settings.

Don't count unique transformed invocations during deduplication; only deduplicate.

FIXME: For some reason, clang doesn't always find initializing declarations for decls that have initializing decls. Clang will instead return the decl itself. This crashes the deduplicator because it will cause us to try to remove the same decl twice. For now I've added a check that if we're removing a var definition, that definition does in fact have an initializing expression. If not, then we just don't remove it. There may be a way to fix this so that we correctly deduplicate these definitions, but for now we'll just have to let some duplicate transformed variable definitions remain in the transformed programs.